+++ src/districtheatingsim/gui/LeafletTab/leaflet/images/layers.png
