+++ src/districtheatingsim/gui/LeafletTab/leaflet/images/marker-icon-2x.png
