+++ src/districtheatingsim/gui/LeafletTab/leaflet/images/marker-icon.png
