+++ src/districtheatingsim/gui/LeafletTab/leaflet/images/marker-shadow.png
